+++ code/imktk/__init__.py
@@ -1,10 +1,15 @@
 import os
+from importlib import metadata
 from .toolkit import _add_folder
 
+__version__ = metadata.version('imktk')
 
 def main():
     loc, da, ds = location()
-    print(f"File location is '{loc}' with folders {da} and {ds} for methods")
+    print(f'  Library version: {__version__}')
+    print(f" Library location: {os.path.dirname(loc)}")
+    print(f"Dataarray scripts: {da}")
+    print(f"  Dataset scripts: {ds}")
 
 
 def location():