+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Docker image file size.
 - Source release file size.
+### Added
+- Version and script location for `imktk`.
 
 ## [0.1.5] - 2022-01-11
 ### Added
-- Changelog.md to repository
+- Changelog.md to repository.